# openapi.yaml
openapi: 3.0.3
info:
  title: JW.org Meeting Content API
  description: |
    Automated API for Jehovah's Witnesses weekly meeting content. This API provides structured data for midweek and weekend meetings, including materials, sections, and media resources.
    
    ## Features
    - Weekly meeting content in JSON format
    - **Comprehensive multilingual support** (Spanish, English, Portuguese, French, German, Italian)
    - Language-specific URLs and terminology based on real JW.org patterns
    - Media resources (videos, images, audio, songs)
    - Meeting sections and time management
    - Historical data access
    - Real-time statistics
    - Language-specific API endpoints
    
    ## Supported Languages
    - 🇪🇸 **Spanish (es)** - español
    - 🇺🇸 **English (en)** - English  
    - 🇧🇷 **Portuguese (pt)** - português
    - 🇫🇷 **French (fr)** - français
    - 🇩🇪 **German (de)** - deutsch
    - 🇮🇹 **Italian (it)** - italiano
    
    ## Language Support
    This API implements realistic language support based on actual JW.org patterns:
    - Different URL paths per language (biblioteca vs library)
    - Localized terminology and content structure
    - Language-specific CSS selectors
    - Proper fallback mechanisms
    - Quality over quantity approach (6 well-supported languages vs 1000+ poorly supported)
    
    ## Rate Limiting
    This API is hosted on GitHub Pages and has no explicit rate limiting, but please be respectful with your usage.
    
    ## CORS
...
    
...
servers:
...
paths:
...
  /languages.json:
    get:
      tags:
        - Metadata
      summary: Get supported languages
      description: |
        Returns information about all supported languages including their configuration,
        terminology, URL patterns, and available API endpoints.
        
        This endpoint provides the definitive list of supported languages and their
        specific characteristics based on real JW.org implementation patterns.
      operationId: getSupportedLanguages
      responses:
        '200':
          description: Supported languages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportedLanguages'
              examples:
                supported_languages:
                  summary: Example supported languages response
                  value:
                    languages:
                      - code: "es"
                        name: "Spanish"
                        nativeName: "Español"
                        flag: "🇪🇸"
                        isDefault: false
                        terminology:
                          workbookName: "Guía de actividades para la reunión Vida y Ministerio Cristiano"
                          watchtowerName: "La Atalaya"
                          meetingTypes:
                            midweek: "Vida y Ministerio Cristiano"
                            weekend: "Reunión de fin de semana"
                        paths:
                          workbook: "guia-actividades-reunion-testigos-jehova"
                          watchtower: "biblioteca/revistas"
                        apiEndpoints:
                          latest: "./latest-es.json"
                          weeks: "./weeks-es.json"
                      - code: "en"
                        name: "English"
                        nativeName: "English"
                        flag: "🇺🇸"
                        isDefault: true
                        terminology:
                          workbookName: "Our Christian Life and Ministry Meeting Workbook"
                          watchtowerName: "The Watchtower"
                          meetingTypes:
                            midweek: "Our Christian Life and Ministry"
                            weekend: "Weekend Meeting"
                        paths:
                          workbook: "jw-meeting-workbook"
                          watchtower: "library/magazines"
                        apiEndpoints:
                          latest: "./latest-en.json"
                          weeks: "./weeks-en.json"
                    meta:
                      endpoint: "languages"
                      description: "Supported languages and their configuration"
                      generatedAt: "2025-08-02T12:00:00.000Z"
                      totalLanguages: 6
                      version: "1.0"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    EndpointMeta:
      type: object
      description: Metadata included in endpoint responses
      required:
        - endpoint
        - description
        - generatedAt
        - version
      properties:
        endpoint:
          type: string
          description: Endpoint name
          example: "latest"
        description:
          type: string
          description: Endpoint description
          example: "Datos de la semana más reciente"
        generatedAt:
          type: string
          format: date-time
          description: When the response was generated
          example: "2025-08-02T00:34:09.741Z"
        version:
          type: string
          description: API version
          example: "1.0"
    SupportedLanguages:
      type: object
      description: Complete list of supported languages with their configuration
      required:
        - languages
        - meta
      properties:
        languages:
          type: array
          description: Array of supported language configurations
          items:
            $ref: '#/components/schemas/LanguageConfig'
        meta:
          $ref: '#/components/schemas/EndpointMeta'
    LanguageConfig:
      type: object
      description: Configuration for a specific language
      required:
        - code
        - name
        - nativeName
        - flag
        - isDefault
        - terminology
        - paths
        - apiEndpoints
      properties:
        code:
          type: string
          description: ISO 639-1 language code
          enum: [es, en, pt, fr, de, it]
          example: es
        name:
          type: string
          description: Language name in English
          example: Spanish
        nativeName:
          type: string
          description: Language name in native language
          example: Español
        flag:
          type: string
          description: Unicode flag emoji
          example: 🇪🇸
        isDefault:
          type: boolean
          description: Whether this is the default language
          example: false
        terminology:
          type: object
          description: Language-specific terminology
          required:
            - workbookName
            - watchtowerName
            - meetingTypes
          properties:
            workbookName:
              type: string
              description: Name of the meeting workbook in this language
              example: "Guía de actividades para la reunión Vida y Ministerio Cristiano"
            watchtowerName:
              type: string
              description: Name of the Watchtower magazine in this language
              example: "La Atalaya"
            meetingTypes:
              type: object
              required:
                - midweek
                - weekend
              properties:
                midweek:
                  type: string
                  example: "Vida y Ministerio Cristiano"
                weekend:
                  type: string
                  example: "Reunión de fin de semana"
        paths:
          type: object
          description: URL path components for this language
          required:
            - workbook
            - watchtower
          properties:
            workbook:
              type: string
              description: URL path component for workbook content
              example: "guia-actividades-reunion-testigos-jehova"
            watchtower:
              type: string
              description: URL path component for watchtower content
              example: "biblioteca/revistas"
        apiEndpoints:
          type: object
          description: Language-specific API endpoints
          required:
            - latest
            - weeks
          properties:
            latest:
              type: string
              description: Latest week endpoint for this language
              example: "./latest-es.json"
            weeks:
              type: string
              description: All weeks endpoint for this language
              example: "./weeks-es.json"
    Error:
      type: object
      description: Standard error response
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
          example: "NotFound"
        message:
          type: string
          description: Human-readable error message
          example: "The requested week data was not found"
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2025-08-02T00:34:09.741Z"
        details:
          type: string
          description: Additional error details
        path:
          type: string
          description: Request path that caused the error
          example: "/data/2025/week-99.json"
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              summary: Internal server error
              value:
                error: "InternalServerError"
                message: "An unexpected error occurred while processing your request"
                timestamp: "2025-08-02T00:34:09.741Z"
  securitySchemes: {}
security: []
...